generator/requirements.txt
diffusers
numpy
opencv-python
requests
Pillow
pytorch_lightning
scikit-image
transformers
torch
flask
flask-cors
accelerate